# o_donjon/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API O'donjon",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://odonjon.fr/",
            "description": "API pour O'donjon"
        }
    ],
    "paths": {
...
        "/campaign/{id}/character": {
            "get": {
                "tags": [
                    "Campaign"
                ],
                "operationId": "App\\Controller\\CampaignController::browseCharacter",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Characters of a campaign",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/campaignCharacters"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/character": {
            "get": {
                "tags": [
                    "Character"
                ],
                "operationId": "App\\Controller\\CharacterController::browse",
                "responses": {
                    "200": {
                        "description": "List of characters of the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/character"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Character"
                ],
                "operationId": "App\\Controller\\CharacterController::add",
                "requestBody": {
                    "$ref": "#/components/requestBodies/characterName"
                },
                "responses": {
                    "200": {
                        "description": "Informations of the character",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/characterInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/character/{id}": {
            "get": {
                "tags": [
                    "Character"
                ],
                "operationId": "App\\Controller\\CharacterController::read",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations of the character",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/characterInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Character"
                ],
                "operationId": "App\\Controller\\CharacterController::edit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/character"
                },
                "responses": {
                    "200": {
                        "description": "Informations of the character",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/characterInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Character"
                ],
                "operationId": "App\\Controller\\CharacterController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "campaignCharacters": {
                "description": "List of Characters of the campaign",
                "properties": {
                    "characters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/character"
                        }
                    }
                },
                "type": "object"
            },
...
            "character": {
                "description": "Information of a character",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Dwalïn",
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "dwalin.png",
                        "nullable": true
                    },
                    "age": {
                        "type": "integer",
                        "default": 0,
                        "example": "97"
                    },
                    "initiative": {
                        "type": "integer",
                        "default": 0,
                        "example": "2"
                    },
                    "height": {
                        "type": "integer",
                        "default": 0,
                        "example": "137"
                    },
                    "weight": {
                        "type": "integer",
                        "default": 0,
                        "example": "61"
                    },
                    "eyes": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "marron",
                        "nullable": true
                    },
                    "skin": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "blanche",
                        "nullable": true
                    },
                    "hair": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "roux",
                        "nullable": true
                    },
                    "appearance": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "personality_traits": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "ideals": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "bonds": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "flaws": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "allies_and_organizations": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "backstory": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "treasure": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "background": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "forgeron",
                        "nullable": true
                    },
                    "alignement": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Loyal neutre",
                        "nullable": true
                    },
                    "attacks_and_spellcasting": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "equipment": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "other_proficiencies_and_languages": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "features_and_traits": {
                        "type": "string",
                        "example": "Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus impedit sequi amet ipsa, recusandae cumque, culpa ullam officia optio qui reiciendis similique, illo illum minus quaerat vitae aliquid iure labore!",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "notFound": {
                "description": "ressource doesn't existe"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID of the ressource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
...
            "characterName": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "Dwalïn",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}